Update LOF model to use include age/sex/bmi and use distinct measurements
